aug 19th incomplete class changes

+++ BedrockBank/Transaction.cs
@@ -18,7 +18,7 @@ namespace BedrockBank
         [Key]
         public int Id { get; set; }
         public string Description { get; set; }
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public TransactionType TypeofTransaction { get; set; }
 
         public DateTime TransactionTime { get; set; }
